requirements.txt
# Core dependencies per ReWTSE-LLM-RL

# Google Gemini API
google-generativeai>=0.3.0

# Machine Learning & RL
torch>=2.0.0
torchvision>=0.15.0
gym>=0.26.0
stable-baselines3>=2.0.0

# Data processing
pandas>=1.5.0
numpy>=1.23.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Financial data
yfinance>=0.2.0

# Optimization (per ReWTSE QP)
cvxopt>=1.3.0

# Progress bars
tqdm>=4.64.0

# Configuration
pyyaml>=6.0

# Optional: API finanziarie aggiuntive
# alpaca-trade-api>=3.0.0
# fredapi>=0.5.0
# sec-api>=1.0.0

# Optional: Technical indicators (richiede compilazione)
# TA-Lib>=0.4.0  # Nota: richiede installazione separata delle librerie C

# Development
jupyter>=1.0.0
ipykernel>=6.19.0
